+++ sonar-magik-plugin/src/main/java/nl/ramsolutions/sw/sonar/MagikSonarWayProfile.java
@@ -6,9 +6,9 @@ import org.sonar.api.server.profile.BuiltInQualityProfilesDefinition;
 import org.sonarsource.analyzer.commons.BuiltInQualityProfileJsonLoader;
 
 /**
- * Magik profile.
+ * Magik Sonar Way profile.
  */
-public final class MagikProfile implements BuiltInQualityProfilesDefinition {
+public final class MagikSonarWayProfile implements BuiltInQualityProfilesDefinition {
 
     static final String PROFILE_NAME = "Sonar way";
     static final String PROFILE_LOCATION = CheckList.PROFILE_DIR + "/Sonar_way_profile.json";
@@ -16,7 +16,7 @@ public final class MagikProfile implements BuiltInQualityProfilesDefinition {
     @Override
     public void define(final Context context) {
         final NewBuiltInQualityProfile profile = context.createBuiltInQualityProfile(PROFILE_NAME, Magik.KEY);
-        BuiltInQualityProfileJsonLoader.load(profile, CheckList.REPOSITORY_KEY, PROFILE_LOCATION);
+        BuiltInQualityProfileJsonLoader.load(profile, CheckList.REPOSITORY_KEY, MagikSonarWayProfile.PROFILE_LOCATION);
         profile.done();
     }